requirements.txt
# L.I.F.E Platform - Python 3.13 Compatible Dependencies
# Azure Marketplace Ready - Enterprise Neural Processing System
# Copyright 2025 - Sergio Paya Borrull

# Azure Functions Runtime (Python 3.13 compatible)
azure-functions>=1.19.0
azure-functions-worker>=1.2.0

# Core Scientific Computing (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Neural Processing & Machine Learning (Python 3.13 compatible)
tensorflow>=2.15.0
torch>=2.1.0
mne>=1.5.0  # EEG/MEG analysis
nilearn>=0.10.2  # Neuroimaging
joblib>=1.3.2

# Azure SDK Components (Latest versions for Python 3.13)
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-monitor-query>=1.1.0
azure-keyvault-secrets>=4.7.0
azure-servicebus>=7.11.0
azure-functions>=1.15.0
azure-cosmos>=4.5.0

# API & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0

# Database & Caching
redis>=4.6.0
pymongo>=4.4.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Monitoring & Logging
opencensus-ext-azure>=1.1.0
opencensus-ext-logging>=0.1.1
structlog>=23.1.0
prometheus-client>=0.17.0

# Security & Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Data Processing & Analysis
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Enterprise Features
celery>=5.3.0  # Task queue
flower>=2.0.0  # Celery monitoring
gunicorn>=21.2.0  # WSGI server
psutil>=5.9.0  # System monitoring

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
typer>=0.9.0

# Production Deployment
docker>=6.1.0
kubernetes>=27.2.0
